Moved webhook actions to API
no refs

Moved webhook actions to API for easier maintenance and consistency with other
API actions
