+++ src/api/index.ts
@@ -5,3 +5,5 @@ export { profileGetFollowersAction } from './action/profile/get-followers';
 export { profileGetFollowingAction } from './action/profile/get-following';
 export { profileGetPostsAction } from './action/profile/get-posts';
 export { searchAction } from './action/search';
+export { webookPostPublishedAction } from './action/webhook/post-published';
+export { webhookSiteChangedAction } from './action/webhook/site-changed';
